+++ src/pdm_conda/conda.py
@@ -1,13 +1,17 @@
+import contextlib
 import json
 import subprocess
 from functools import lru_cache
 from tempfile import NamedTemporaryFile
+from typing import Iterable
 from urllib.parse import urlparse
 
+from packaging.version import Version
 from pdm.exceptions import RequirementError
 from pdm.models.setup import Setup
 
 from pdm_conda.models.candidates import CondaCandidate
+from pdm_conda.models.config import CondaRunner
 from pdm_conda.models.requirements import (
     CondaRequirement,
     Requirement,
@@ -19,6 +23,15 @@ from pdm_conda.project import CondaProject
 from pdm_conda.utils import normalize_name
 
 
+@contextlib.contextmanager
+def _optional_temporary_file(environment: dict):
+    if environment:
+        with NamedTemporaryFile(mode="w+", suffix=".yml") as f:
+            yield f
+    else:
+        yield
+
+
 def run_conda(cmd, **environment) -> dict:
     """
     Creates temporary environment file and run conda command
@@ -26,7 +39,7 @@ def run_conda(cmd, **environment) -> dict:
     :param environment: environment data
     :return: conda command response
     """
-    with NamedTemporaryFile(mode="w+", suffix=".yml") as f:
+    with _optional_temporary_file(environment) as f:
         if environment:
             for name, options in environment.items():
                 if options:
@@ -36,29 +49,71 @@ def run_conda(cmd, **environment) -> dict:
             f.seek(0)
             cmd = cmd + ["-f", f.name]
         process = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, encoding="utf-8")
-        if "--json" in cmd:
-            try:
-                response = json.loads(process.stdout)
-            except:
-                response = {}
-        else:
-            response = {"message": process.stdout}
+    if "--json" in cmd:
         try:
-            process.check_returncode()
-        except subprocess.CalledProcessError:
-            msg = "Error locking dependencies\n"
-            if isinstance(response, dict):
-                if response.get("success", False):
-                    pass
-                if err := response.get("solver_problems", []):
-                    msg += "\n".join(err)
-                elif err := response.get("message", []):
-                    msg += err
-            else:
-                msg += process.stderr
-            msg += str(environment)
-            raise RequirementError(msg)
+            response = json.loads(process.stdout)
+        except:
+            response = {}
+    else:
+        response = {"message": process.stdout}
+    try:
+        process.check_returncode()
+    except subprocess.CalledProcessError as e:
+        msg = "Error locking dependencies\n"
+        if isinstance(response, dict) and not response.get("success", False):
+            if err := response.get("solver_problems", response.get("error", response.get("message", []))):
+                if isinstance(err, str):
+                    err = [err]
+                msg += "\n".join(err)
+        else:
+            msg += process.stderr
+        if environment:
+            msg += f"\n{environment}"
+        raise RequirementError(msg) from e
     return response
+
+
+def _sort_packages(packages: list[dict]) -> Iterable[dict]:
+    """
+    Sort packages following mamba specification
+    (https://mamba.readthedocs.io/en/latest/advanced_usage/package_resolution.html).
+    :param packages: list of conda packages
+    :return: sorted conda packages
+    """
+    if len(packages) <= 1:
+        return packages
+    _with_track_features = []
+    _sorted = []
+    _by_version: dict[str, list] = dict()
+
+    # get most recent version with timestamp
+    for p in packages:
+        _by_version.setdefault(p["version"], []).append(p)
+
+    name = packages[0].get("name", "")
+    # sort by version
+    for _, pkgs in sorted(
+        _by_version.items(),
+        key=lambda x: Version(parse_conda_version(x[0], inverse=name != "openssl")),
+    ):
+        # sort same version packages using build number
+        if len(pkgs) > 1:
+            _by_build_number: dict[int, list] = dict()
+            for p in pkgs:
+                _by_build_number.setdefault(p.get("build_number", 0), []).append(p)
+            pkgs.clear()
+            # sort same build number by timestamp
+            for n in sorted(_by_build_number.keys()):
+                pkgs.extend(sorted(_by_build_number[n], key=lambda p: p.get("timestamp", 0)))
+
+        # track_feature to bottom
+        for p in pkgs:
+            if p.get("track_feature", ""):
+                _with_track_features.append(p)
+            else:
+                _sorted.append(p)
+
+    return reversed(_with_track_features + _sorted)
 
 
 @lru_cache(maxsize=None)
@@ -74,29 +129,33 @@ def _conda_search(
     :param channels: requirement channels
     :return: list of conda candidates
     """
-    command = project.conda_config.command("search")
+    config = project.conda_config
+    command = config.command("search")
     if not project.virtual_packages:
         project.virtual_packages = conda_virtual_packages(project)
 
     command.append(requirement)
     for c in channels:
         command.extend(["-c", c])
+    if channels:
+        command.append("--override-channels")
     command.append("--json")
-    result = run_conda(command)
+    try:
+        result = run_conda(command)
+    except RequirementError as e:
+        if "PackagesNotFoundError:" in str(e):
+            result = dict()
+        else:
+            raise
+
     candidates = []
     # sort values per build number (greater first)
-    packages = result.get("result", dict()).get("pkgs", [])
-    _packages: dict[tuple, list] = dict()
-    for p in packages:
-        name = p["name"]
-        version = p["version"]
-        k = (name, version)
-        _packages.setdefault(k, []).append(p)
-    packages = []
-    for ps in _packages.values():
-        packages.extend(sorted(reversed(ps), key=lambda p: p["build_number"], reverse=True))
+    if config.runner != CondaRunner.MICROMAMBA:
+        packages = result.get(parse_requirement(f"conda:{requirement}").name, [])
+    else:
+        packages = result.get("result", dict()).get("pkgs", [])
 
-    for p in packages:
+    for p in _sort_packages(packages):
         dependencies = p.get("depends", [])
         valid_candidate = True
         for d in dependencies:
@@ -136,7 +195,7 @@ def conda_search(
         requirement = requirement.as_line(with_build_string=True, conda_compatible=True).replace(" ", "=")
     if "::" in requirement:
         channel, requirement = requirement.split("::", maxsplit=1)
-    channels = [channel] if channel else (project.conda_config.channels or ["defaults"])
+    channels = [channel] if channel else project.conda_config.channels
     return _conda_search(requirement, project, tuple(channels))
 
 
@@ -200,7 +259,7 @@ def conda_install(
         command.append("--dry-run")
     if config.installation_method == "copy":
         _copy = "copy"
-        if config.runner == "micromamba":
+        if config.runner == CondaRunner.MICROMAMBA:
             _copy = f"always-{_copy}"
         command.append(f"--{_copy}")
 
@@ -226,7 +285,7 @@ def conda_uninstall(
     command = config.command("remove")
     if no_deps:
         _prune = "no-prune"
-        if config.runner == "conda":
+        if config.runner != CondaRunner.MICROMAMBA:
             _prune = "force-remove"
         command.append(f"--{_prune}")
     if dry_run:
@@ -249,8 +308,12 @@ def conda_virtual_packages(project: CondaProject) -> set[CondaRequirement]:
     virtual_packages = set()
     if config.is_initialized:
         info = run_conda(config.command("info") + ["--json"])
-        # todo: fix conda
-        virtual_packages = {parse_requirement(f"conda:{p}") for p in info["virtual packages"]}
+        if config.runner != CondaRunner.MICROMAMBA:
+            _virtual_packages = {"=".join(p) for p in info["virtual_pkgs"]}
+        else:
+            _virtual_packages = set(info["virtual packages"])
+
+        virtual_packages = {parse_requirement(f"conda:{p.replace('=', '==', 1)}") for p in _virtual_packages}
     return virtual_packages