+++ src/pdm_conda/models/candidates.py
@@ -141,7 +141,8 @@ class CondaCandidate(Candidate):
         url = package["url"]
         for k, v in hashes.items():
             url += f"#{k}={v}"
-        name, version, build_string = package["name"], package["version"], package.get("build_string", "")
+        name, version = package["name"], package["version"]
+        build_string = package.get("build", package.get("build_string", ""))
         channel = package["channel"]
         req = parse_requirement(f"conda:{name} {version} {build_string}")
         req.is_python_package = requires_python is not None