+++ src/pdm_conda/models/config.py
@@ -1,6 +1,7 @@
 import os
 from contextlib import contextmanager
 from dataclasses import dataclass, field
+from enum import Enum
 from functools import wraps
 from pathlib import Path
 from typing import Any
@@ -51,6 +52,12 @@ def is_conda_config_initialized(project: Project):
     return "conda" in project.pyproject.settings
 
 
+class CondaRunner(str, Enum):
+    CONDA = "conda"
+    MAMBA = "mamba"
+    MICROMAMBA = "micromamba"
+
+
 @dataclass
 class PluginConfig:
     _project: Project = field(repr=False, default=None)
@@ -68,11 +75,15 @@ class PluginConfig:
     mapping_download_dir: Path = field(repr=False, default=Path())
 
     def __post_init__(self):
-        if self.runner not in ["conda", "micromamba", "mamba"]:
+        if self.runner not in list(CondaRunner):
             raise ProjectError(f"Invalid Conda runner: {self.runner}")
         if self.installation_method not in ["hard-link", "copy"]:
             raise ProjectError(f"Invalid Conda installation method: {self.installation_method}")
-        to_suscribe = [(self._project.pyproject._data, "update"), (self._project.pyproject, "reload")]
+        to_suscribe = [
+            (self._project.pyproject._data, "update"),
+            (self._project.pyproject, "write"),
+            (self._project.pyproject, "reload"),
+        ]
         for obj, name in to_suscribe:
             func = getattr(obj, name)
             if not is_decorated(func):
@@ -162,7 +173,9 @@ class PluginConfig:
         :param cmd: command, install by default
         :return: args list
         """
-        _command = [self.runner, cmd, "-y"]
-        if cmd in ("install", "create") or (cmd == "search" and self.runner != "conda"):
+        _command = [self.runner, cmd]
+        if cmd in ("install", "create", "remove"):
+            _command.append("-y")
+        if cmd in ("install", "create") or (cmd == "search" and self.runner == CondaRunner.MICROMAMBA):
             _command.append("--strict-channel-priority")
         return _command